Added terrain but doesn't work correctly, added GameObject Fence in Unity (but script is empty)
