+++ PF_TaxiDriver/Assets/New Terrain.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3bd1bb01081b467478949605990d13d2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
