+++ PF_TaxiDriver/Assets/Scripts/Fence.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c666b3c81db18d44cb767d1a128276df
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
