--- PF_TaxiDriver/Assets/Scripts/Taxi.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 73e7ae9898fad604e810a9f404a1d9ae
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
